remove sort.py
