+++ python/ray/data/_internal/arrow_ops/transform_pyarrow.py
@@ -6,7 +6,7 @@ except ImportError:
     pyarrow = None
 
 if TYPE_CHECKING:
-    from ray.data._internal.sort import SortKey
+    from ray.data._internal.planner.exchange.sort_task_spec import SortKey
 
 
 def sort(table: "pyarrow.Table", sort_key: "SortKey") -> "pyarrow.Table":
