+++ python/ray/data/_internal/pandas_block.py
@@ -33,7 +33,7 @@ if TYPE_CHECKING:
     import pandas
     import pyarrow
 
-    from ray.data._internal.sort import SortKey
+    from ray.data._internal.planner.exchange.sort_task_spec import SortKey
     from ray.data.aggregate import AggregateFn
 
 T = TypeVar("T")
