+++ python/ray/data/_internal/planner/exchange/aggregate_task_spec.py
@@ -1,7 +1,7 @@
 from typing import List, Optional, Tuple, Union
 
 from ray.data._internal.planner.exchange.interfaces import ExchangeTaskSpec
-from ray.data._internal.sort import SortKey
+from ray.data._internal.planner.exchange.sort_task_spec import SortKey
 from ray.data._internal.table_block import TableBlockAccessor
 from ray.data.aggregate import AggregateFn, Count
 from ray.data.aggregate._aggregate import _AggregateOnKeyBase
