+++ python/ray/data/_internal/table_block.py
@@ -11,7 +11,7 @@ from ray.data.block import Block, BlockAccessor
 from ray.data.row import TableRow
 
 if TYPE_CHECKING:
-    from ray.data._internal.sort import SortKey
+    from ray.data._internal.planner.exchange.sort_task_spec import SortKey
 
 
 T = TypeVar("T")
